+++ src/server/bitops_family.cc
@@ -5,7 +5,6 @@
 #include <absl/strings/ascii.h>
 #include <absl/strings/match.h>
 
-#include <bitset>
 #include <nonstd/expected.hpp>
 
 #include "base/logging.h"
@@ -16,6 +15,7 @@
 #include "server/command_families.h"
 #include "server/command_registry.h"
 #include "server/conn_context.h"
+#include "server/db_slice.h"
 #include "server/engine_shard_set.h"
 #include "server/error.h"
 #include "server/namespaces.h"
