+++ src/server/dflycmd.cc
@@ -23,7 +23,6 @@
 #include "facade/dragonfly_listener.h"
 #include "facade/reply_builder.h"
 #include "server/cluster_support.h"
-#include "server/debugcmd.h"
 #include "server/engine_shard_set.h"
 #include "server/error.h"
 #include "server/journal/journal.h"
@@ -31,7 +30,7 @@
 #include "server/main_service.h"
 #include "server/namespaces.h"
 #include "server/rdb_save.h"
-#include "server/script_mgr.h"
+#include "server/replica.h"
 #include "server/server_family.h"
 #include "server/server_state.h"
 #include "server/transaction.h"
