+++ src/server/namespaces.cc
@@ -6,6 +6,8 @@
 
 #include "base/flags.h"
 #include "base/logging.h"
+#include "server/blocking_controller.h"
+#include "server/db_slice.h"
 #include "server/engine_shard_set.h"
 
 ABSL_DECLARE_FLAG(bool, cache_mode);
