+++ src/server/server_family.cc
@@ -63,9 +63,9 @@ extern "C" {
 #include "server/memory_cmd.h"
 #include "server/multi_command_squasher.h"
 #include "server/namespaces.h"
-#include "server/protocol_client.h"
 #include "server/rdb_load.h"
 #include "server/rdb_save.h"
+#include "server/replica.h"
 #include "server/script_mgr.h"
 #include "server/search/search_family.h"
 #include "server/server_state.h"
