+++ src/server/server_state.cc
@@ -6,8 +6,6 @@
 
 #include <mimalloc.h>
 
-#include "server/acl/user_registry.h"
-
 extern "C" {
 #include "redis/zmalloc.h"
 }
@@ -17,7 +15,7 @@ extern "C" {
 #include "base/logging.h"
 #include "facade/conn_context.h"
 #include "facade/dragonfly_connection.h"
-#include "server/channel_store.h"
+#include "facade/facade_stats.h"
 #include "server/common.h"
 #include "server/journal/journal.h"
 #include "util/listener_interface.h"
