+++ src/server/snapshot.h
@@ -6,22 +6,23 @@
 
 #include <deque>
 
-#include "base/pod_array.h"
-#include "core/search/base.h"
-#include "io/file.h"
 #include "server/db_slice.h"
-#include "server/execution_state.h"
 #include "server/rdb_save.h"
 #include "server/synchronization.h"
 #include "server/table.h"
 #include "server/tiered_storage.h"
-#include "util/fibers/future.h"
 
 namespace dfly {
+
+class ExecutionState;
 
 namespace journal {
 struct Entry;
 }  // namespace journal
+
+namespace search {
+using DocId = uint32_t;
+}  // namespace search
 
 // ┌────────────────┐   ┌─────────────┐
 // │IterateBucketsFb│   │  OnDbChange │
